Added rotated effects for chain guns

+++ core/src/io/anuke/ucore/function/SegmentConsumer.java
@@ -1,5 +1,5 @@
 package io.anuke.ucore.function;
 
 public interface SegmentConsumer{
-	public void accept(int x, int y, int x2, int y2);
+	public void accept(float x, float y, float x2, float y2);
 }
